The playgrid plugin is a mess right now.
The bulk of the logic is in keyring/includes/services/extended/playgrid.php, which needs to be moved into the playgrid plugin.
The OAuth2 flow is working against PGP, creating a WP_User, logging them in and redirecting them to their profile page.

Still can't initiate the OAuth2 flow for a non-authenticated user.